Support unix datagram socket
The helper can use now a unix datagram socket instead of a file
descriptor.

If the --socket argument is specified, the helper enters a new
connection phase. The helper create a unix datagram socket at the
specified path, and waits until the first packet arrives.

When receiving the first packet, the helper connects the socket to the
client address, to allow sendmsg_x() and write() to send packets to the
client, and ensure we never read any packet send by another client.

Existing clients such as vfkit and krunkit send an invalid "handshake"
packet ("VFKT"). The helper drop the invalid first packet (len < 64)
since forwarding it to vmnet fails.

Once connected to the client, the helper continues with the normal flow.
At this point a passed file descriptor or connected unix socket should
be the same.