Add --list-shared-interfaces option
The option prints interface names available for use with
VMNET_BRIDGED_MODE. This will allow selecting shared interface
automatically, and checking if bridged mode can be tested in a developer
machines.

In github runners the option reports a shared interface, but the vm does
not get an IP address when using bridged mode.