Show all tests in same plot
Collect the results from all tests, and plot them using one barh() per
test. This shows all test in the same scale, making it easy to
understand the performance of the test configurations in all tests.

The socket_vmnet plot was separated to vz and qemu plots, since showing
6 configurations for 3 tests is too much. This makes it harder to
compare vz and qemu performance, but the main point of that plot is
how adding vms decrease the performance.