Generalize image helpers
Use disk["image"] to limit the number of temporary variable, since we
need to add similar code to create kernel and initrd images.

Make create_image() more generic so we can reuse it for creating kernel
and initrd images that do not need any conversion.

Extract clone() helper to make the code nicer to work with. We will need
to clone also the kernel and initrd images.