Add optional image properties
Add "kernel", "kernel_parameters", and "initrd" properties to image spec.
If kernel and initrd are specified the urls are downloaded to the cache
and the data is cloned to the vm directory. If kernel_parameters is
specified it is copied to the disk dict.

If the optional disk properties are specified we use them in the qemu
command. This will be used to debug the random kernel panics in the
github runner.

We can use these properties later for vfkit and krunkit it needed.