Skip jq install if possible
This minimize the time to start the VM in the CI, making it easier to
debug startup issue, when we don't have ssh access to the guest. The
time moves to the next step preparing the VM, but it is easier to debug
that step with ssh access.

Before:
- start vm: 2m15s
- prepare vm: 15s

After:
- start vm: 1m33s
- prepare vm: 1m1s