Introduce vmnet-client
The *vmnet-client* program makes it possible to start a VM using *QEMU*
or *vfkit* from a shell script.

The client accepts a *QEMU* or *vfkit* command configured to use file
descriptor 4 as connected unix datagram socket. The client creates a
socketpair, starts `vmnet-helper` with one socket, and starts the
provided command with the other socket.

This will help users of *socket_vmnet_client*[1] to migrate to
*vmnet-helper*.

The *vfkit* command can be used without *vmnet-client* with the
`--socket` option. With *vmnet-client* you can use *vfkit* from the
shell without creating a bound unix socket.

The current version supports only vment shared mode and uses a random
interface-id for every run. We can add command line options later to
support more *vmnet-helper* features.

Apple recommends to use the MAC address returned by
vmnet_start_interface() in the vmnet_mac_address key but testing show
that this is not required. Using the recommended MAC address requires
parsing *vmnet-helper* standard output and modifying the user command to
use the returned MAC address. This makes it more complicated to
implement and use the program so we keep it simple and let the user
configure the MAC address.

[1] https://github.com/lima-vm/socket_vmnet/tree/master?tab=readme-ov-file#qemu