Adapt example cloud-init for tests
To test the examples we need to detect that VM IP address. Add a runcmd
script reporting the IP address to the serial console.

The script depends on the interface name, rename the interface to vmnet0
to make this more clear.

We run create-iso.sh for every boot to generate a new iso with new
instance id. This forces the runcmd to run on every boot.

Both vfkit and qemu example scripts wait until the guest writes the IP
address to the serial log, and write the IP address to ip-address file.
The tests will use this file to provision the vm.

With this change we can add integration tests for the examples, ensuring
their correctness. The downside is making the examples more complicated.