Expose vment.CLIENT path
Extract helper.PREFIX, helper.HELPER, and helper.CLIENT. The CLIENT is
exposed as part of the vmnet package we can use it for client
integration.

The prefix and executable name are set by meson. Ideally we would
generate the python files so changing in meson.build will update the
entire project. I did not find a good way to do this yet.