Add --connection client to example
In this mode we connect the vm to the vmnet network using the
vmnet-client. Since running the vm via a client is basically passing the
vm command line to the client, the easier way to do this is to is to let
the vm manage the process.

The vm accepts now an optional client argument. When set, starting and
stopping changes:

- The client is injected into the vm command.
- The vm is started in a new process group, including the vmnet-helper
  child process.
- When stopping the vm we terminate the vm process group.
- The vm stdout are redirected to the vm log.

Example run:

    % ./example server --connection client
    Creating cloud-init iso '/Users/nir/.vmnet-helper/vms/server/cidata.iso'
    Starting 'vfkit' virtual machine 'server' with mac address '52:e4:25:c1:63:51'
    Virtual machine IP address: 192.168.105.80