client: Support usable helper options
The client accepts now all the vmnet-helper options that can be used
with the client. We parse and validate the options so we can fail early
without starting the vm command if the helper options are invalid. Valid
options are appended to the helper command.

The --enable-tso and --enable-checksum-offload can be used with the
client, but they make sense only with krunkit. Since krunkit cannot be
used with the client, they are not exposed in the client.

Example usage:

    % build/vmnet-client -h

    Run command with vmnet-helper

        vmnet-client [--interface-id UUID] [--operation-mode shared|bridged|host]
                     [--start-address ADDR] [--end-address ADDR]
                     [--subnet-mask MASK] [--shared-interface NAME]
                     [--enable-isolation] [-v|--verbose] [--version] [-h|--help]
                     -- command ...

With this change we can use all vment-helper features.