docs: Document user and group sudoers rules
If a user is not in the staff group, they cannot use the default sudoers
rule. An admin can solve this by adding a sudoers rule for a specific
user, or creating a new group for vment, adding the user to the group,
and adding the sudoers rule for the group.