Add --enable-virtio-header option
If enabled and running on macOS >= 15.4 we add the
vmnet_enable_virtio_header_key to the interface description.

The documentation[1] does not tell us anything about the purpose of this
option, except except incompatibility with
vment_enable_checksum_offload, so we reject this option when using
--enable-checksum-offload.

To keep it simple, the option is always reported in the online help, but
it is ignored if running on macOS < 15.4. Ignoring is easier for users
since it avoids the need to check the current macOS version.

If the option is enabled we log a new message:

    [main] enabling virtio header

The new option can be enabled in the example scripts for testing.

When enabling this option getting vm ip address times out and we get no
output in serial.log. Maybe this option requires special kernel
configuration.

[1] https://developer.apple.com/documentation/vmnet/vmnet_enable_virtio_header_key-swift.var?language=objc