example: Rename --vmnet-offload to --enable-offloading
Matching other boolean flags and the option name in krunkit. Update the
benchmarks to use "offloading" instead of "offload" and use a boolean
option instead of string ("on", "off").

Since krunkit supports now testing with and without offloading, update
the benchmark to test with system krunkit instead of krunkit.local.

Update the plots to compare krunkit with and without offloading. Since
the default is not using offloading, change the tag to "offload".