build: Build universal binary
meson does not support well building universal binaries so we use a
build script to build arm64 and x86_64 binaries, create universal
binaries, install them to a temporary directory, and create an archive
with the built executables.

Local development can be done using meson commands, but in the CI we use
the build script.

Building a release tarball with universal binaries:

    ./build.sh build

Build directory structure:

    % tree -L1 build
    build
    ├── arm64
    ├── root
    ├── vmnet-client
    ├── vmnet-helper
    ├── vmnet-helper.tar.gz
    └── x86_64