docs: Simplify installation section
Modify the description to match the new one-line install.

Add a description of the install process, since we don't show the actual
installation commands.

Eliminate the "granting permission" section since the script handles
this part now. Add link to [sudoers.d](sudoers.d) for users that want to
learn more about this.