client: Pass only socket fd to children
Previously we passed both helper fd (3) and command fd (4) to both the
helper and the command. This is not needed an can prevent the helper
from detecting closing of the command socket. In practice this is
harmless since the helper cannot detect closing the command socket
anyway with datagram socket.

Now we close the unneeded socket after forking, so the child get only
the socket it needs.