helper: Fix detection of closed socket
If we fallback to read() we did not handle zero read correctly. We would
return 1 packet instead of 0 packets. This breaks closing the connection
on zero read.

In practice this does not help since we never get a zero read when the
peer close the socket, and writing an empty packet to vment will fail
and terminate the helper. Fixed anyway for clarity.