client: Terminate helper when command terminates
Previously the client started the helper and replaced itself with the
command. When the command terminates, the helper is left running, since
it cannot detect that the command closed its socket. Users do not have a
good way to terminate the helper and they should not be responsible for
terminating the helper since they did not start it.

Fixed by starting the helper and the command, and waiting for the
command. When the command terminates for any reason, the helper is
terminated as well.

To simplify usage, the user can terminate the client process, and the
client terminates both the command and the helper.

The client exit code is the exit code of the command, keeping the
previous behaviour when the client was replaced by the command.

Termination is very simple:
- We send a SIGTERM signal and wait for the child process. If the
  child does not terminate the client will wait forever.
- The helper is terminated after terminating the command. This should
  not be an issue since it terminates very quickly.