Support for running as unprivileged user
Turns out that on macOS 26 the helper does not need root if signed with
"com.apple.security.virtualization" entitlement.

Add a `codesign` step to the build script. Signing the fat binary
applies the signature to all architectures. This has does not help on
older macOS 26 but is harmless, so we keep the build script simple and
always sign the executable.

Add --unprivileged option to the example test to allow running as
unprivileged user to allow manual testing on macOS 26. We cannot use
this in the CI yet since we don't have macOS 26 intel runners.

Change the helper to drop privileges only if running as root, and log
the effective uid an gid on start and after dropping privileges.

Example helper log when starting as root:

    INFO  [main] running /opt/vmnet-helper/bin/vmnet-helper unknown on macOS 15.7.3
    INFO  [main] running as uid: 0 gid: 0
    INFO  [main] enabling bulk forwarding
    DEBUG [main] starting vmnet interface
    INFO  [main] started vmnet interface
    INFO  [main] dropped privileges to uid: 501 gid: 20
    ...

This change is enough for creating a pre-release for testing, but to
complete the change we need more work:

- Update the client with similar option
- Update the install script to skip the sudoers configuration when
  running on macOS 26.